+++ src/api/routes/notification.rs
@@ -3,7 +3,7 @@
 //! The routes are integrated with the Actix-web framework and interact with the notification controller.
 
 use crate::{
-    api::controllers::notifications,
+    api::controllers::notification,
     models::{
         DefaultAppState, NotificationCreateRequest, NotificationUpdateRequest, PaginationQuery,
     },
@@ -16,7 +16,7 @@ async fn list_notifications(
     query: web::Query<PaginationQuery>,
     data: web::ThinData<DefaultAppState>,
 ) -> impl Responder {
-    notifications::list_notifications(query.into_inner(), data).await
+    notification::list_notifications(query.into_inner(), data).await
 }
 
 /// Retrieves details of a specific notification by ID.
@@ -25,7 +25,7 @@ async fn get_notification(
     notification_id: web::Path<String>,
     data: web::ThinData<DefaultAppState>,
 ) -> impl Responder {
-    notifications::get_notification(notification_id.into_inner(), data).await
+    notification::get_notification(notification_id.into_inner(), data).await
 }
 
 /// Creates a new notification.
@@ -34,7 +34,7 @@ async fn create_notification(
     request: web::Json<NotificationCreateRequest>,
     data: web::ThinData<DefaultAppState>,
 ) -> impl Responder {
-    notifications::create_notification(request.into_inner(), data).await
+    notification::create_notification(request.into_inner(), data).await
 }
 
 /// Updates an existing notification.
@@ -44,7 +44,7 @@ async fn update_notification(
     request: web::Json<NotificationUpdateRequest>,
     data: web::ThinData<DefaultAppState>,
 ) -> impl Responder {
-    notifications::update_notification(notification_id.into_inner(), request.into_inner(), data)
+    notification::update_notification(notification_id.into_inner(), request.into_inner(), data)
         .await
 }
 
@@ -54,7 +54,7 @@ async fn delete_notification(
     notification_id: web::Path<String>,
     data: web::ThinData<DefaultAppState>,
 ) -> impl Responder {
-    notifications::delete_notification(notification_id.into_inner(), data).await
+    notification::delete_notification(notification_id.into_inner(), data).await
 }
 
 /// Configures the notification routes.
